+++ telegrambots-meta/src/main/java/org/telegram/telegrambots/api/objects/payments/ShippingQuery.java
@@ -1,0 +1,57 @@
+package org.telegram.telegrambots.api.objects.payments;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import org.telegram.telegrambots.api.interfaces.BotApiObject;
+import org.telegram.telegrambots.api.objects.User;
+
+/**
+ * @author Ruben Bermudez
+ * @version 1.0
+ *
+ * This object contains information about incoming shipping query.
+ */
+public class ShippingQuery implements BotApiObject {
+    private static final String ID_FIELD = "id";
+    private static final String FROM_FIELD = "from";
+    private static final String INVOICE_PAYLOAD_FIELD = "invoice_payload";
+    private static final String SHIPPING_ADDRESS_FIELD = "shipping_address";
+
+    @JsonProperty(ID_FIELD)
+    private String id; ///< Unique query identifier
+    @JsonProperty(FROM_FIELD)
+    private User from; ///< User who sent the query
+    @JsonProperty(INVOICE_PAYLOAD_FIELD)
+    private String invoicePayload; ///< Bot specified invoice payload
+    @JsonProperty(SHIPPING_ADDRESS_FIELD)
+    private ShippingAddress shippingAddress; ///< User specified shipping address
+
+    public ShippingQuery() {
+        super();
+    }
+
+    public String getId() {
+        return id;
+    }
+
+    public User getFrom() {
+        return from;
+    }
+
+    public String getInvoicePayload() {
+        return invoicePayload;
+    }
+
+    public ShippingAddress getShippingAddress() {
+        return shippingAddress;
+    }
+
+    @Override
+    public String toString() {
+        return "ShippingQuery{" +
+                "id='" + id + '\'' +
+                ", from=" + from +
+                ", invoicePayload='" + invoicePayload + '\'' +
+                ", shippingAddress=" + shippingAddress +
+                '}';
+    }
+}